+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Add Wine Staging again to now be used alongside Wine GE.
 - Change Arch Linux repositories to use newer upstream packages than SteamOS.
 - Change installation of SteamOS to optionally work on Arch Linux and Manjaro live media.
+- Change RAM mounts to use 'tmpfs' instead of 'ramfs' for FUSE compatibility.
 
 ## 3.0.1